pyMQTT-Ingest/requirements.txt
paho-mqtt==1.3.0
psutil==5.4.1
